Hooks for package version collection
