+++ benchmark/qutip/runbenchmark.py
@@ -1,20 +1,5 @@
 import sys
 import pytest
-from importlib.metadata import version, PackageNotFoundError
-
-def pkg_version(package):
-    try:
-        return version(package)
-    except PackageNotFoundError:
-        return None
-
-@pytest.hookimpl()
-def pytest_benchmark_update_json(config, benchmarks, output_json):
-    output_json["package_versions"]["scipy"] = pkg_version("scipy")
-    output_json["package_versions"]["numpy"] = pkg_version("numpy")
-    output_json["package_versions"]["cython"] = pkg_version("cython")
-    output_json["package_versions"]["qutip"] = pkg_version("qutip")
-    output_json["package_versions"]["pytest"] = pkg_version("pytest")
 
 def run_benchmarks():
     return pytest.main(
@@ -32,7 +17,6 @@ def run_benchmarks():
 
 def main():
     return run_benchmarks()
-
 
 if __name__ == "__main__":
     sys.exit(main())
